Model Runner Sanity unittests
Former-commit-id: 4180c45837ab141b666dd8abc0bf6f90b7ef5a32

+++ streamline/tests/test_zclean.py
@@ -1,5 +1,7 @@
 import shutil
+DEBUG = False
 
 
 def test_stub():
-    shutil.rmtree('./tests/')
+    if not DEBUG:
+        shutil.rmtree('./tests/')
